# Conflicts:
#	lib/filesystem.py

+++ lib/filesystem.py
@@ -24,7 +24,7 @@ import charliecloud as ch
 # To see the directory formats in released versions:
 #
 #   $ git grep -F 'STORAGE_VERSION =' $(git tag | sort -V)
-STORAGE_VERSION = 5
+STORAGE_VERSION = 6
 
 
 ## Globals ##
@@ -430,7 +430,10 @@ class Path(pathlib.PosixPath):
          ch.FATAL("can’t symlink: %s -> %s: %s" % (self.name, target,
                                                    x.strerror))
 
-   def unlink_(self, *args, **kwargs):
+   def unlink_(self, missing_ok=False):
+      # FIXME: Once we require Python 3.8, we can just pass through missing_ok.
+      if (missing_ok and not self.exists_()):
+         return
       ch.ossafe(super().unlink, "can’t unlink: %s" % self.name)
 
 
@@ -542,7 +545,7 @@ class Storage:
       if (v_found == STORAGE_VERSION):
          ch.VERBOSE("found storage dir v%d: %s" % (STORAGE_VERSION, self.root))
          self.lock()
-      elif (v_found in {None, 2, 3, 4}):  # initialize/upgrade
+      elif (v_found in {None, 2, 3, 4, 5}):  # initialize/upgrade
          ch.INFO("%s storage directory: v%d %s"
                  % (op, STORAGE_VERSION, self.root))
          self.root.mkdir_()
@@ -555,12 +558,23 @@ class Storage:
          self.build_large.mkdir_()
          self.unpack_base.mkdir_()
          self.upload_cache.mkdir_()
-         for old in self.unpack_base.iterdir():
-            new = old.parent // str(old.name).replace(":", "+")
-            if (old != new):
-               if (new.exists()):
-                  ch.FATAL("can’t upgrade: already exists: %s" % new)
-               old.rename(new)
+         if (v_found is not None):  # upgrade
+            if (v_found < 3):
+               # Escape colon in image names as plus starting in v3.
+               for old in self.unpack_base.iterdir():
+                  new = old.parent // str(old.name).replace(":", "+")
+                  if (old != new):
+                     if (new.exists()):
+                        ch.FATAL("can’t upgrade: already exists: %s" % new)
+                     old.rename(new)
+            if (v_found < 6):
+               # Git metadata moved from /.git to /ch/.git in v6.
+               for img in self.unpack_base.iterdir():
+                  old = img // ".git"
+                  new = img // "ch/git"
+                  if (old.exists_()):
+                     new.parent.mkdir_()
+                     old.rename_(new)
          self.version_file.file_write("%d\n" % STORAGE_VERSION)
       else:                         # can’t upgrade
          ch.FATAL("incompatible storage directory v%d: %s"
@@ -686,7 +700,7 @@ class TarFile(tarfile.TarFile):
    def add_(self, name, **kwargs):
       def filter_(ti):
          assert (ti.name == "." or ti.name[:2] == "./")
-         if (ti.name.startswith("./.git") or ti.name == "./ch/git.pickle"):
+         if (ti.name in ("./ch/git", "./ch/git.pickle")):
             ch.DEBUG("omitting from push: %s" % ti.name)
             return None
          self.fix_member_uidgid(ti)
